+++ packages/kiarina-lib-google-auth/src/kiarina/lib/google/auth/settings.py
@@ -13,6 +13,34 @@ class GoogleAuthSettings(BaseSettings):
     type: Literal["default", "service_account", "user_account"] = "default"
 
     # --------------------------------------------------
+    # Fields (common)
+    # --------------------------------------------------
+
+    project_id: str | None = None
+
+    impersonate_service_account: str | None = None
+    """
+    Email address of the service account to impersonate
+
+    The source principal requires the roles/iam.serviceAccountTokenCreator role.
+    Note that this required permission is not included in the roles/owner role.
+    """
+
+    scopes: list[str] = Field(
+        default_factory=lambda: [
+            "https://www.googleapis.com/auth/cloud-platform",  # All GCP resources
+            "https://www.googleapis.com/auth/drive",  # Google Drive resources
+            "https://www.googleapis.com/auth/spreadsheets",  # Google Sheets resources
+        ]
+    )
+    """
+    List of scopes to request during the authentication
+
+    Specify the scopes required for impersonation authentication.
+    Specify the scopes required for user account authentication.
+    """
+
+    # --------------------------------------------------
     # Fields (service_account)
     # --------------------------------------------------
 
@@ -47,34 +75,6 @@ class GoogleAuthSettings(BaseSettings):
     If cache is not available, use for refresh.
     Refreshed credentials will be cached.
     """
-
-    # --------------------------------------------------
-    # Fields (common)
-    # --------------------------------------------------
-
-    impersonate_service_account: str | None = None
-    """
-    Email address of the service account to impersonate
-
-    The source principal requires the roles/iam.serviceAccountTokenCreator role.
-    Note that this required permission is not included in the roles/owner role.
-    """
-
-    scopes: list[str] = Field(
-        default_factory=lambda: [
-            "https://www.googleapis.com/auth/cloud-platform",  # All GCP resources
-            "https://www.googleapis.com/auth/drive",  # Google Drive resources
-            "https://www.googleapis.com/auth/spreadsheets",  # Google Sheets resources
-        ]
-    )
-    """
-    List of scopes to request during the authentication
-
-    Specify the scopes required for impersonation authentication.
-    Specify the scopes required for user account authentication.
-    """
-
-    project_id: str | None = None
 
     # --------------------------------------------------
     # Validators