+++ packages/kiarina-lib-google-auth/tests/test_get_default_credentials.py
@@ -4,6 +4,7 @@ import google.oauth2.credentials
 import google.oauth2.service_account
 import pytest
 
+from kiarina.lib.google.auth import settings_manager
 from kiarina.lib.google.auth._utils.get_default_credentials import (
     get_default_credentials,
 )
@@ -21,15 +22,12 @@ def test_adc():
     assert isinstance(credentials, google.oauth2.credentials.Credentials)
 
 
-@pytest.mark.xfail(
-    "KIARINA_LIB_GOOGLE_AUTH_TEST_GCP_SA_KEY_FILE" not in os.environ,
-    reason="GCP SA key file not set",
-)
-def test_service_account():
-    # Set the environment variable to point to a service account key file
-    os.environ["GOOGLE_APPLICATION_CREDENTIALS"] = os.path.expanduser(
-        os.environ["KIARINA_LIB_GOOGLE_AUTH_TEST_GCP_SA_KEY_FILE"]
-    )
+def test_service_account(load_settings, monkeypatch):
+    if "GOOGLE_APPLICATION_CREDENTIALS" not in os.environ:
+        settings = settings_manager.get_settings_by_key("service_account_file")
+        monkeypatch.setenv(
+            "GOOGLE_APPLICATION_CREDENTIALS", settings.service_account_file
+        )
 
     credentials = get_default_credentials()
     print(f"Obtained service account credentials of type: {type(credentials)}")