+++ packages/kiarina-lib-google-auth/tests/test_get_self_signed_jwt.py
@@ -1,22 +1,9 @@
-import os
-
-import pytest
-
-from kiarina.lib.google.auth import GoogleAuthSettings, get_self_signed_jwt
+from kiarina.lib.google.auth import get_self_signed_jwt
 
 
-@pytest.mark.xfail(
-    "KIARINA_LIB_GOOGLE_AUTH_TEST_GCP_SA_KEY_FILE" not in os.environ,
-    reason="GCP SA key file not set",
-)
-def test_get_self_signed_jwt():
+def test_get_self_signed_jwt(load_settings):
     jwt = get_self_signed_jwt(
-        settings=GoogleAuthSettings(
-            type="service_account",
-            service_account_file=os.environ[
-                "KIARINA_LIB_GOOGLE_AUTH_TEST_GCP_SA_KEY_FILE"
-            ],
-        ),
+        "service_account_file",
         audience="https://blazeworks.jp/",
     )
     assert jwt.count(".") == 2